backend/interview-service/requirements.txt
fastapi
uvicorn
pyautogen
google-generativeai
pydantic
python-dotenv
firebase-admin
